requirements.txt
requests>=2.31.0
openpyxl>=3.1.2
PyYAML>=6.0.1
jinja2>=3.1.3
